requirements.txt
pyTelegramBotAPI==4.8.0
python-dotenv==0.19.2
requests==2.28.1
python-telegram-bot-calendar==1.0.5
python-telegram-bot==13.15
loguru==0.6.0
translators==5.5.5
peewee==3.15.4